component폴더를 components로 변경
